content/057/services/order/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
asyncpg==0.29.0
aio-pika==9.4.3
pydantic==2.8.2
opentelemetry-distro==0.47b0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-instrumentation-asyncpg==0.47b0
opentelemetry-instrumentation-aio-pika==0.47b0
pytest==8.3.2
pytest-asyncio==0.23.8
